[alloc] Create a new API for free lists
Very simple and lightweight... :)
The simple collector is broken right now

Should we abstract behind a trait?
